refactor (image) : redoc 관련 어노테이션 추가 및 api 별 response 분리
